Refactoring of compatible collections
--HG--
extra : convert_revision : charlie%40nunit.com-20091102145446-sr910o07zjrs34fe

+++ src/framework/Shared/Framework/Constraints/EqualConstraint.cs
@@ -416,7 +416,7 @@ namespace NUnit.Framework.Constraints
         /// <param name="depth">The depth of this failure in a set of nested collections</param>
         private void DisplayCollectionDifferences(MessageWriter writer, ICollection expected, ICollection actual, int depth)
         {
-            long failurePoint = comparer.FailurePoints.Count > depth ? (long)comparer.FailurePoints[depth] : -1;
+            int failurePoint = comparer.FailurePoints.Count > depth ? (int)comparer.FailurePoints[depth] : -1;
 
             DisplayCollectionTypesAndSizes(writer, expected, actual, depth);
